refactor(server): rename pipeline_jobs.py → jobs.py (#1101 chunk 1 sub-1)
First sub-commit of Epic #1101 chunk 1 (foundation). Pure refactor;
no behaviour change. Isolating the rename in its own commit makes
the chunk bisectable and the subsequent enrichment-job-type
additions clean.

Rationale: the file already serves the pipeline job queue + spawn +
registry abstractions; chunk 1 adds COMMAND_ENRICHMENT alongside the
existing COMMAND_FULL = "full_incremental_pipeline". The promote-
queued / cancel / stale-reconcile / pid-alive logic is
command_type-agnostic. Calling the file pipeline_jobs.py while it
serves enrichment jobs too would be misleading.

Decision: O4 in docs/wip/RFC-088-CHUNK1-LOCK-AUDIT.md.

Changes:

  src/:
    pipeline_jobs.py → jobs.py (new module docstring)
    + import updates in 3 files:
      - jobs_log_path.py
      - scheduler.py
      - routes/jobs.py (HTTP route module — different package, no
        clash with the engine module; isort autofix applied)
    + docstring update in pipeline_run_prometheus.py (the Prometheus
      metric name "podcast_pipeline_jobs_finished_total" is a stable
      public interface consumed by Grafana dashboards — left
      intentionally unchanged)

  tests/:
    test_pipeline_jobs_argv.py → test_jobs_argv.py
    test_pipeline_jobs_helpers.py → test_jobs_helpers.py
    test_pipeline_jobs_log_pump.py → test_jobs_log_pump.py
    test_pipeline_jobs_reconcile.py → test_jobs_reconcile.py
    + their import statements updated

  docs:
    + 4 markdown references updated (RFC-079, RFC-098,
      SERVER-SIDE-GAP-ANALYSIS, RFC-088 plan)

Validation:
  - 536 tests/integration/server/ pass green.
  - 35 renamed tests pass green (full module surface exercised).
  - make docs strict mkdocs build: green.

Next sub-commit: foundation types (protocol.py, envelope.py, paths.py,
correlation.py, scorers/protocol.py).
