+++ tests/integration/server/test_jobs_log_pump.py
@@ -8,7 +8,7 @@ from typing import Any, cast
 
 import pytest
 
-from podcast_scraper.server.pipeline_jobs import (
+from podcast_scraper.server.jobs import (
     _LOG_MAX_BYTES_DEFAULT,
     _pump_subprocess_to_log,
     job_log_max_bytes,
